adding scripts and moving files around
