+++ scripts/prodDeployAndroid.sh
@@ -1,0 +1,4 @@
+cd android
+./gradlew assembleRelease
+cd ..
+NODE_ENV=production react-native run-android --variant=release
