+++ src/App.js
@@ -2,18 +2,18 @@ import React from 'react';
 import { createStackNavigator } from 'react-navigation';
 import { ApolloProvider } from 'react-apollo';
 import { YellowBox } from 'react-native';
-import { client } from './src/config';
+import { client } from './config';
 
-import { StartScreen } from './src/pages/start/main';
-import { HomeScreen } from './src/pages/home/main';
-import { ChoresScreen } from './src/pages/chores/main';
-import { AddChoreScreen } from './src/pages/add/chore/main';
-import { IouScreen } from './src/pages/ious/main';
-import { PayScreen } from './src/pages/ious/pay';
-import { SplitCostScreen } from './src/pages/ious/split';
-import { AllTransaction } from './src/pages/ious/allTrans';
-import { ShoppingListScreen } from './src/pages/items/shoppinglist';
-import { AddToShoppingListScreen } from './src/pages/items/addToShoppingList';
+import { StartScreen } from './pages/start/main';
+import { HomeScreen } from './pages/home/main';
+import { ChoresScreen } from './pages/chores/main';
+import { AddChoreScreen } from './pages/add/chore/main';
+import { IouScreen } from './pages/ious/main';
+import { PayScreen } from './pages/ious/pay';
+import { SplitCostScreen } from './pages/ious/split';
+import { AllTransaction } from './pages/ious/allTrans';
+import { ShoppingListScreen } from './pages/items/shoppinglist';
+import { AddToShoppingListScreen } from './pages/items/addToShoppingList';
 
 YellowBox.ignoreWarnings(['Warning: isMounted(...) is deprecated', 'Module RCTImageLoader']);
